requirements.txt
django
djangorestframework
bs4
requests
httpx
parsel
django-cors-headers
numpy
pandas
matplotlib
cmake
cvxpy
django-rest-passwordreset
sendgrid
surprise
fuzzywuzzy
django-environ
psycopg2-binary
sqlparse
stripe
dj-database-url
django-heroku
gunicorn
whitenoise
django4-background-tasks